Use user-unit dir from systemd pkg-config file
This commit drops the debian/patches/system-not-required.patch
since it no longer applies. It also drops the duplication of
the user-unit dir in debian/rules in favour of just picking it
up from systemd.pc. This means we need a build-dep on systemd
which ships the pkg-config file we want to read from.
The package already only targets linux systems, and even if it
did not AFAIK there's a shim pkg-config file available on non-linux
to avoid having to special-case them. I thus don't see any reason
why we'd need to patch things out and duplicate information.