debian/patches: Ensure support with nvidia drivers >= 495
the NVIDIA >= 495 drivers apparently allows creating and importing dma-bufs.
However, this won't succeed, but without an error.

The result is that undefined content is drawn onto the screen on clients.
This affects both the RDP and VNC backend.

Disabling dma-bufs in such case gets rid of this issue, because with this
patch g-r-d always assumes, that dma-bufs are unavailable with the NVIDIA
driver.