Manually depend on systemd-sysusers & systemd-tmpfiles
It's not worth attempting to tweak postinst to handle
the non-systemd use case ourselves. If someone needs
that, they can file a bug or a merge request.

The functionality could be manually done in postinst
similar to how it was done in polkitd.postinst.

Arguably, that could be handled by dh_installtmpfiles
and dh_installsysusers.

But gnome-control-center, gdm3, etc. simply are not
designed to work outside systemd.

Even patches are problematic since it's unlikely
that someone is going to be testing and maintaining
those code paths